requirements.txt
google-generativeai
pandas
pyyaml
jinja2
